Backend/pyproject.toml
[project]
name = "saas-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.16.5",
    "celery>=5.5.3",
    "chromadb>=1.0.20",
    "cryptography>=45.0.6",
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-redis>=0.1.2",
    "mem0ai>=0.1.116",
    "pandas>=2.3.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pypdf>=6.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis[async]>=6.4.0",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
]
